+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.4.0-rc.25",
+  sdkVersion: "3.4.0-rc.26",
   genVersion: "2.578.0",
-  userAgent: "speakeasy-sdk/typescript 3.4.0-rc.25 2.578.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.4.0-rc.26 2.578.0 2.0 @orq-ai/node",
 } as const;